pyproject.toml
[project]
name = "data-analyst-agent"
version = "0.1.0"
description = "Data Analyst Agent API using FastAPI and Google Gemini"
authors = [
    {name = "Ishan", email = "23f1000733@ds.study.iitm.ac.in"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "google-genai>=0.3.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "duckdb>=0.9.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "plotly>=5.17.0",
    "Pillow>=10.0.0",
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0",
    "lxml>=4.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
[tool.hatch.build.targets.wheel]
include = ["main.py"]
